{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "New Project"
version = "0.1.0"
description = "Super new project"
authors = ["Author"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
fastapi = "^0.112"
uvicorn = {extras = ["standard"], version = "^0.30"}
django-split-settings = "^1.1.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21"}
aerich = "^0.7.1"
httpx = "^0.27"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^7.1.3"
requests = "^2.28.1"

[tool.aerich]
tortoise_orm = "config.DB_CONFIG"
location = "app/db/migrations"
src_folder = "./app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
